test(RoundResultTest): 테스트를 클래스에 맞게 재배치
-test: 테스트를 위한 MockParticipant 생성
