+++ src/test/java/domain/RoundHistoryTest.java
@@ -12,6 +12,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import mock.MockParticipant;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -56,6 +57,41 @@ class RoundHistoryTest {
       //then
       Map<RoundResult, Integer> expected = new HashMap<>(Map.of(RoundResult.WIN, 1));
       assertThat(roundHistory.getDealerResult()).isEqualTo(expected);
+    }
+
+    @Nested
+    @DisplayName("라운드 결과를 반환한다.")
+    class Round {
+
+      @Test
+      @DisplayName("올바르게 승자를 가려낸다.")
+      void test_round() {
+        // given
+        final Participant winner = new MockParticipant();
+        winner.hit(new TrumpCard(Rank.ACE, Suit.CLUB));
+
+        final Participant loser = new MockParticipant();
+        loser.hit(new TrumpCard(Rank.TEN, Suit.CLUB));
+        // when&then
+        assertThat(RoundResult.round(winner, loser)).isEqualTo(RoundResult.WIN);
+        assertThat(RoundResult.round(loser, winner)).isEqualTo(RoundResult.LOSE);
+      }
+
+      @Test
+      @DisplayName("21점이 넘는다면, 상대방이 승리한다.")
+      void test_RoundOverBurst() {
+        // given
+        final Participant buster = new MockParticipant();
+        buster.hit(new TrumpCard(Rank.TEN, Suit.CLUB));
+        buster.hit(new TrumpCard(Rank.TEN, Suit.CLUB));
+        buster.hit(new TrumpCard(Rank.TEN, Suit.CLUB));
+
+        final Participant player = new MockParticipant();
+        player.hit(new TrumpCard(Rank.TEN, Suit.CLUB));
+        // when&then
+        assertThat(RoundResult.round(player, buster)).isEqualTo(RoundResult.WIN);
+        assertThat(RoundResult.round(buster, player)).isEqualTo(RoundResult.LOSE);
+      }
     }
   }
 }
